2.1.8 巡河轨迹优化，湖泊的公示牌以及建议功能（湖泊相关暂时隐藏）

+++ hZSZ/src/main/java/com/zju/rchz/model/Lake.java
@@ -5,7 +5,7 @@ package com.zju.rchz.model;
  */
 
 public class Lake {
-    public long lakeId;
+    public int lakeId;
     public String lakeName;//湖泊名称
     public int lakeLevel;//等级
     public int waterType;
